Move CreateRaceResponder in accurate folder
